Added support for forward type references
