+++ gutta-apievolution-dsl/src/main/java/gutta/apievolution/dsl/ApiRevisionModelBuilderPass2.java
@@ -5,9 +5,8 @@ import gutta.apievolution.dsl.parser.ApiRevisionBaseVisitor;
 import gutta.apievolution.dsl.parser.ApiRevisionParser;
 import org.antlr.v4.runtime.Token;
 
-import java.util.*;
+import java.util.Optional;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import static gutta.apievolution.dsl.parser.ApiRevisionLexer.*;
 
@@ -21,11 +20,10 @@ import static gutta.apievolution.dsl.parser.ApiRevisionLexer.*;
  * @param <S> The concrete service type
  * @param <O> The concrete service operation type
  */
-abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends RecordType<A, R, F>,
+abstract class ApiRevisionModelBuilderPass2<A extends ApiDefinition<A>, R extends RecordType<A, R, F>,
         F extends Field<R, F>, E extends EnumType<A, E, M>, M extends EnumMember<E, M>,
-        S extends Service<A, S, O, R>, O extends ServiceOperation<S, O, R>> extends ApiRevisionBaseVisitor<Void> {
-
-    private int currentTypeId;
+        S extends Service<A, S, O, R>, O extends ServiceOperation<S, O, R>>
+        extends ApiRevisionModelBuilderPass<A, R, F, E, M, S, O> {
 
     protected A currentRevision;
 
@@ -37,54 +35,34 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
 
     protected S currentService;
 
-    protected A buildRevision(final ApiRevisionParser.ApiDefinitionContext apiRevisionSpec,
+    protected void augmentRevision(final ApiRevisionParser.ApiDefinitionContext apiRevisionSpec,
+                              A apiDefinition,
                               final Optional<A> optionalPredecessor) {
-        this.currentTypeId = 0;
-        this.currentRevision = null;
+        this.currentRevision = apiDefinition;
         this.currentRecordType = null;
         this.previousRevision = optionalPredecessor;
 
         apiRevisionSpec.accept(this);
-
-        return this.currentRevision;
-    }
-
-    private int getNextTypeId() {
-        return this.currentTypeId++;
     }
 
     @Override
     public final Void visitApiDefinition(final ApiRevisionParser.ApiDefinitionContext ctx) {
-        QualifiedName name = this.buildQualifiedName(ctx.name);
-
-        this.currentRevision = this.createRevision(ctx, name, this.handleAnnotations(ctx.annotations),
-                this.previousRevision);
-
         // Process sub elements
         ctx.elements.forEach(element -> element.accept(this));
 
         return null;
     }
 
-    protected abstract A createRevision(ApiRevisionParser.ApiDefinitionContext context, QualifiedName name,
-                                        Set<Annotation> annotations, Optional<A> predecessor);
-
     @Override
     public final Void visitRecordType(final ApiRevisionParser.RecordTypeContext ctx) {
-        String name = this.identifierAsText(ctx.name);
-
-        // Determine internal name
-        Optional<String> internalName = this.determineInternalName(ctx.as);
-
-        // Handle modifiers, if any
-        RecordTypeModifiers modifiers = this.determineModifiers(ctx);
+        String typeName = this.identifierAsText(ctx.name);
+        R recordType = this.resolveRecord(typeName, ctx.name.start);
 
         // Resolve super type, if applicable
         Optional<String> superTypeName = this.optionalIdentifierAsText(ctx.superType);
-        Optional<R> superType = this.resolveRecord(superTypeName, () -> ctx.superType.start);
+        Optional<R> optionalSuperType = this.resolveRecord(superTypeName, () -> ctx.superType.start);
+        optionalSuperType.ifPresent(superType -> recordType.setSuperType(superType));
 
-        R recordType = this.createRecordType(ctx, name, internalName, this.getNextTypeId(), this.currentRevision,
-                modifiers.isAbstract(), superType);
         this.registerNewRecordType(recordType);
 
         this.currentRecordType = recordType;
@@ -94,61 +72,21 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
         return null;
     }
 
-    protected abstract R createRecordType(final ApiRevisionParser.RecordTypeContext context, final String name,
-                                          final Optional<String> internalName, int typeId, final A currentRevision,
-                                          final boolean abstractFlag, Optional<R> superType);
+    private R resolveRecord(String name, Token nameToken) {
+        Optional<UserDefinedType<A>> optionalType = this.currentRevision.resolveUserDefinedType(name);
 
-    protected void registerNewRecordType(final R recordType) {
-        // Do nothing by default
-    }
-
-    private static final String MSG_ONLY_ONE_OPTIONALITY_MODIFIER = "Only one optionality modifier may be specified.";
-
-    private RecordTypeModifiers determineModifiers(final ApiRevisionParser.RecordTypeContext context) {
-        Optional<Boolean> abstractFlag = Optional.empty();
-        Optional<Optionality> optionality = Optional.empty();
-
-        // Process given modifiers
-        for (ApiRevisionParser.RecordModifierContext modifierContext : context.modifiers) {
-            Token modifierToken = modifierContext.start;
-
-            switch (modifierToken.getType()) {
-                case K_ABSTRACT:
-                    abstractFlag = setOnce(abstractFlag, Boolean.TRUE, modifierToken,
-                            () -> "Abstract modifier may only be specified once.");
-                    break;
-
-                case K_OPTIONAL:
-                    optionality = setOnce(optionality, Optionality.OPTIONAL, modifierToken,
-                            () -> MSG_ONLY_ONE_OPTIONALITY_MODIFIER);
-                    break;
-
-                case K_OPTIN:
-                    optionality = setOnce(optionality, Optionality.OPT_IN, modifierToken,
-                            () -> MSG_ONLY_ONE_OPTIONALITY_MODIFIER);
-                    break;
-
-                case K_MANDATORY:
-                    optionality = setOnce(optionality, Optionality.MANDATORY, modifierToken,
-                            () -> MSG_ONLY_ONE_OPTIONALITY_MODIFIER);
-                    break;
-
-                default:
-                    throw new IllegalArgumentException("Unsupported modifier type: " + modifierToken.getText());
-            }
+        if (!optionalType.isPresent()) {
+            throw new APIResolutionException(nameToken, "No record type named '" + name + "'.");
         }
 
-        // Use defaults, if necessary
-        return new RecordTypeModifiers(abstractFlag.orElse(Boolean.FALSE), optionality.orElse(Optionality.MANDATORY));
-    }
+        R type = this.assertRecordType(optionalType.get());
 
-    private static <T> Optional<T> setOnce(final Optional<T> currentValue, final T newValue, final Token referenceToken,
-                                           final Supplier<String> errorMessageSupplier) {
-        if (currentValue.isPresent()) {
-            throw new APIParseException(referenceToken, errorMessageSupplier.get());
-        } else {
-            return Optional.of(newValue);
+        if (type == null) {
+            throw new APIResolutionException(nameToken, "User-defined type '" + name +
+                    "' is not a record type.");
         }
+
+        return type;
     }
 
     private Optional<R> resolveRecord(final Optional<String> name, final Supplier<Token> nameToken) {
@@ -192,10 +130,6 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
                                      final Optional<String> internalName, Type type, Optionality optionality,
                                      final R owner);
 
-    protected void registerNewField(final F field) {
-        // Do nothing by default
-    }
-
     private Optionality determineFieldModifiers(final ApiRevisionParser.FieldModifierContext context) {
         if (context == null) {
             return Optionality.MANDATORY;
@@ -218,12 +152,9 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
 
     @Override
     public final Void visitEnumType(final ApiRevisionParser.EnumTypeContext ctx) {
-        String name = this.identifierAsText(ctx.name);
+        String typeName = this.identifierAsText(ctx.name);
+        E enumType = this.resolveEnumType(typeName, ctx.name.start);
 
-        // Determine internal name
-        Optional<String> internalName = this.determineInternalName(ctx.as);
-
-        E enumType = this.createEnumType(ctx, name, internalName, this.getNextTypeId(), this.currentRevision);
         this.registerNewEnumType(enumType);
 
         this.currentEnumType = enumType;
@@ -233,11 +164,21 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
         return null;
     }
 
-    protected abstract E createEnumType(ApiRevisionParser.EnumTypeContext context, String name,
-                                        Optional<String> internalName, int typeId, A owner);
+    private E resolveEnumType(String name, Token nameToken) {
+        Optional<UserDefinedType<A>> optionalType = this.currentRevision.resolveUserDefinedType(name);
 
-    protected void registerNewEnumType(final E enumType) {
-        // Do nothing by default
+        if (!optionalType.isPresent()) {
+            throw new APIResolutionException(nameToken, "No enum type named '" + name + "'.");
+        }
+
+        E type = this.assertEnumType(optionalType.get());
+
+        if (type == null) {
+            throw new APIResolutionException(nameToken, "User-defined type '" + name +
+                    "' is not a enum type.");
+        }
+
+        return type;
     }
 
     @Override
@@ -256,16 +197,12 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
     protected abstract M createEnumMember(ApiRevisionParser.EnumMemberContext context, String name,
                                           Optional<String> internalName, E owner);
 
-    protected void registerNewEnumMember(final M enumMember) {
-        // Do nothing by default
-    }
-
     @Override
     public final Void visitService(final ApiRevisionParser.ServiceContext ctx) {
-        String name = this.identifierAsText(ctx.name);
+        String name = identifierAsText(ctx.name);
 
         // Determine internal name
-        Optional<String> internalName = this.determineInternalName(ctx.as);
+        Optional<String> internalName = determineInternalName(ctx.as);
 
         S service = this.createService(ctx, name, internalName, this.currentRevision);
         this.registerNewService(service);
@@ -279,10 +216,6 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
 
     protected abstract S createService(ApiRevisionParser.ServiceContext context, String name,
                                        Optional<String> internalName, A owner);
-
-    protected void registerNewService(final S service) {
-        // Do nothing by default
-    }
 
     @Override
     @SuppressWarnings("unchecked")
@@ -306,67 +239,6 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
     protected abstract O createServiceOperation(ApiRevisionParser.ServiceOperationContext context, String name,
                                                 Optional<String> internalName, S owner, R returnType, R parameterType);
 
-    protected void registerNewServiceOperation(final O serviceOperation) {
-        // Do nothing by default
-    }
-
-    protected Set<Annotation> handleAnnotations(final List<ApiRevisionParser.AnnotationContext> annotationContexts) {
-        Set<Annotation> annotations = new HashSet<>(annotationContexts.size());
-
-        for (ApiRevisionParser.AnnotationContext annotationContext : annotationContexts) {
-            String typeText = annotationContext.typeToken.getText();
-            String valueText = annotationContext.value.getText();
-
-            // Remove leading '@'
-            String typeName = typeText.substring(1);
-            Annotation annotation = new Annotation(typeName, valueText);
-            annotations.add(annotation);
-        }
-
-        return annotations;
-    }
-
-    protected static String unquote(final String input) {
-        int endIndex = (input.length() - 2);
-
-        return input.substring(1, endIndex);
-    }
-
-    protected List<String> splitQualifiedName(final ApiRevisionParser.QualifiedNameContext context) {
-        return context.parts.stream()
-                .map(this::identifierAsText)
-                .collect(Collectors.toList());
-    }
-
-    private QualifiedName buildQualifiedName(final ApiRevisionParser.QualifiedNameContext context) {
-        List<String> parts = this.splitQualifiedName(context);
-        return new QualifiedName(parts);
-    }
-
-    protected String identifierAsText(final ApiRevisionParser.IdentifierContext context) {
-        if (context.id != null) {
-            return context.id.getText();
-        } else {
-            return unquote(context.literal.getText());
-        }
-    }
-
-    protected Optional<String> optionalIdentifierAsText(final ApiRevisionParser.IdentifierContext context) {
-        if (context == null) {
-            return Optional.empty();
-        } else {
-            return Optional.of(this.identifierAsText(context));
-        }
-    }
-
-    private Optional<String> determineInternalName(final ApiRevisionParser.AsClauseContext context) {
-        if (context == null) {
-            return Optional.empty();
-        } else {
-            return Optional.of(this.identifierAsText(context.aliasName));
-        }
-    }
-
     private Type resolveType(final ApiRevisionParser.TypeReferenceContext context) {
         return new TypeResolver().visit(context);
     }
@@ -377,26 +249,7 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
 
     protected abstract R assertRecordType(UserDefinedType<A> type);
 
-    private static class RecordTypeModifiers {
-
-        private final boolean abstractFlag;
-
-        private final Optionality optionality;
-
-        public RecordTypeModifiers(final boolean abstractFlag, final Optionality optionality) {
-            this.abstractFlag = abstractFlag;
-            this.optionality = optionality;
-        }
-
-        public boolean isAbstract() {
-            return this.abstractFlag;
-        }
-
-        public Optionality getOptionality() {
-            return this.optionality;
-        }
-
-    }
+    protected abstract E assertEnumType(UserDefinedType<A> type);
 
     private class TypeResolver extends ApiRevisionBaseVisitor<Type> {
 
@@ -454,9 +307,9 @@ abstract class ApiRevisionModelBuilder<A extends ApiDefinition<A>, R extends Rec
 
         @Override
         public Type visitUserDefinedTypeReference(final ApiRevisionParser.UserDefinedTypeReferenceContext ctx) {
-            String typeName = ApiRevisionModelBuilder.this.identifierAsText(ctx.typeName);
+            String typeName = ApiRevisionModelBuilderPass2.this.identifierAsText(ctx.typeName);
             Optional<UserDefinedType<A>> optionalType =
-                    ApiRevisionModelBuilder.this.currentRevision.resolveUserDefinedType(typeName);
+                    ApiRevisionModelBuilderPass2.this.currentRevision.resolveUserDefinedType(typeName);
 
             if (optionalType.isPresent()) {
                 return optionalType.get();
