+++ gutta-apievolution-dsl/src/main/java/gutta/apievolution/dsl/ConsumerApiLoader.java
@@ -20,7 +20,14 @@ public class ConsumerApiLoader extends ApiDefinitionLoader {
     public static ConsumerApiDefinition loadFromStream(InputStream inputStream, int referencedRevision) {
         try {
             ApiRevisionParser.ApiDefinitionContext specification = parseStream(inputStream);
-            return new ConsumerApiRevisionModelBuilder().buildConsumerRevision(referencedRevision, specification);
+
+            ConsumerApiRevisionModelBuilderPass1 pass1 = new ConsumerApiRevisionModelBuilderPass1();
+            ConsumerApiRevisionModelBuilderPass2 pass2 = new ConsumerApiRevisionModelBuilderPass2();
+
+            ConsumerApiDefinition apiDefinition = pass1.buildConsumerRevision(specification, referencedRevision);
+            pass2.augmentConsumerRevision(specification, apiDefinition);
+
+            return apiDefinition;
         } catch (IOException e) {
             throw new ApiLoadException("Error loading API definition.", e);
         }
