+++ gutta-apievolution-dsl/src/main/java/gutta/apievolution/dsl/PredecessorType.java
@@ -1,0 +1,7 @@
+package gutta.apievolution.dsl;
+
+enum PredecessorType {
+    NONE,
+    IMPLICIT,
+    EXPLICIT
+}
